Split validators into modules
