+++ tests/test_unit/test_validators.py
@@ -3,11 +3,11 @@ from contextlib import nullcontext as does_not_raise
 import numpy as np
 import pytest
 
-from movement.io.validators import (
+from movement.io.validators.datasets import ValidPosesDataset
+from movement.io.validators.files import (
     ValidDeepLabCutCSV,
     ValidFile,
     ValidHDF5,
-    ValidPosesDataset,
 )
